#29 renamed utils_test to testing_utilities
